requirements.txt
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
google-genai>=1.0.0
Pillow>=10.0.0
streamlit>=1.38.0
python-dotenv>=1.0.0
